flare-streamlit-demo/requirements.txt
streamlit
python-dotenv
cassandra-driver
langchain
openai
cassio
tiktoken